% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{big_epa_cars}
\alias{big_epa_cars}
\title{big_epa_cars}
\format{Tibble with 41,804 features and 83 fields}
\usage{
big_epa_cars
}
\description{
Dataset from tidytuesday: rfordatascience/tidytuesday
}
\keyword{datasets}
