% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataviz.R
\name{mod_dataviz_ui}
\alias{mod_dataviz_ui}
\alias{mod_dataviz_server}
\title{mod_dataviz_ui and mod_dataviz_server}
\usage{
mod_dataviz_ui(id, type = c("point", "hist", "boxplot", "bar"))

mod_dataviz_server(input, output, session, type)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
