% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_raw.R
\name{mod_raw_ui}
\alias{mod_raw_ui}
\alias{mod_raw_server}
\title{mod_raw_ui and mod_raw_server}
\usage{
mod_raw_ui(id)

mod_raw_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
